+++ engine/runtime/src/main/java/org/enso/compiler/TreeToIr.java
@@ -1,5 +1,6 @@
 package org.enso.compiler;
 
+import java.util.ArrayList;
 import java.util.UUID;
 import org.enso.compiler.core.IR;
 import org.enso.compiler.core.IR$Application$Literal$Sequence;
@@ -22,6 +23,7 @@ import org.enso.compiler.core.IR$Error$Syntax$EmptyParentheses$;
 import org.enso.compiler.core.IR$Error$Syntax$InvalidImport$;
 import org.enso.compiler.core.IR$Error$Syntax$Reason;
 import org.enso.compiler.core.IR$Error$Syntax$UnrecognizedToken$;
+import org.enso.compiler.core.IR$Error$Syntax$UnsupportedSyntax;
 import org.enso.compiler.core.IR$Expression$Binding;
 import org.enso.compiler.core.IR$Expression$Block;
 import org.enso.compiler.core.IR$Foreign$Definition;
@@ -87,6 +89,76 @@ final class TreeToIr {
     */
   IR.Module translate(Tree ast) {
     return translateModule(ast);
+  }
+
+  /**
+   * Translates an inline program expression represented in the parser {@link Tree}
+   * to the compiler's {@link IR} representation.
+   *
+   * Inline expressions must <b>only</b> be expressions, and may not contain any
+   * type of definition.
+   *
+   * @param ast The tree representing the expression to translate.
+   * @return The {@link IR} representation of the given ast if it is valid, otherwise
+   *  {@link Option#empty()}.
+   */
+  Option<IR.Expression> translateInline(Tree ast) {
+    return switch(ast) {
+      case Tree.BodyBlock b -> {
+        List<IR.Expression> expressions = nil();
+        java.util.List<IR.IdentifiedLocation> locations = new ArrayList<>();
+        for (Line statement : b.getStatements()) {
+          Tree exprTree = statement.getExpression();
+          IR.Expression expr = switch (exprTree) {
+            case Tree.Export x -> null;
+            case Tree.Import x -> null;
+            case Tree.Invalid x -> null;
+            case null -> null;
+            default -> translateExpression(exprTree);
+          };
+          if (expr != null) {
+            expressions = cons(expr, expressions);
+            if (expr.location().isDefined()) {
+              locations.add(expr.location().get());
+            }
+          }
+        }
+        if (expressions.size() == 0) {
+          yield Option.empty();
+        } else if (expressions.size() == 1) {
+          yield Option.apply(expressions.apply(0));
+        } else {
+          Option<IdentifiedLocation> combinedLocation;
+          if (locations.isEmpty()) {
+            combinedLocation = Option.empty();
+          } else {
+            combinedLocation = Option.apply(
+                new IdentifiedLocation(
+                    new Location(
+                      locations.get(1).start(),
+                      locations.get(locations.size() - 1).end()
+                    ),
+                    Option.empty()
+                )
+            );
+          }
+
+          yield Option.apply(
+              new IR$Expression$Block(
+                  expressions,
+                  expressions.last(),
+                  combinedLocation,
+                  false,
+                  null,
+                  null
+              )
+          );
+        }
+      }
+      default -> {
+        throw new IllegalStateException();
+      }
+    };
   }
 
   /** Translate a top-level Enso module into [[IR]].
@@ -168,7 +240,8 @@ final class TreeToIr {
         var body = translateExpression(fn.getBody());
 
         if (body == null) {
-            throw new NullPointerException();
+            var error = translateSyntaxError(inputAst, new IR$Error$Syntax$UnsupportedSyntax("Block without body"));
+            yield cons(error, appendTo);
         }
         var binding = new IR$Module$Scope$Definition$Method$Binding(
           methodRef,
@@ -563,6 +636,9 @@ final class TreeToIr {
           case "->" -> {
             // Old-style lambdas; this syntax will be eliminated after the parser transition is complete.
             var arg = app.getLhs();
+            if (arg == null) {
+              yield translateSyntaxError(app, IR$Error$Syntax$UnexpectedExpression$.MODULE$);
+            }
             var isSuspended = new boolean[1];
             if (arg instanceof Tree.UnaryOprApp susApp && "~".equals(susApp.getOpr().codeRepr())) {
                 arg = susApp.getRhs();
@@ -1140,6 +1216,18 @@ final class TreeToIr {
         new IR$Pattern$Literal(translateLiteral(lit), getIdentifiedLocation(lit), meta(), diag());
       case Tree.Number num ->
         new IR$Pattern$Literal((IR.Literal) translateNumber(num), getIdentifiedLocation(num), meta(), diag());
+      case Tree.UnaryOprApp num when num.getOpr().codeRepr().equals("-") -> {
+        var n = (IR$Literal$Number) translateExpression(num.getRhs());
+        var t = n.copy(
+          n.copy$default$1(),
+          "-" + n.copy$default$2(),
+          n.copy$default$3(),
+          n.copy$default$4(),
+          n.copy$default$5(),
+          n.copy$default$6()
+        );
+        yield new IR$Pattern$Literal(t, getIdentifiedLocation(num), meta(), diag());
+      }
       case Tree.TypeAnnotated anno -> {
         var type = buildNameOrQualifiedName(maybeManyParensed(anno.getType()));
         var expr = buildNameOrQualifiedName(maybeManyParensed(anno.getExpression()));