+++ engine/runtime/src/test/scala/org/enso/interpreter/test/semantic/DataflowErrorsTest.scala
@@ -204,8 +204,8 @@ class DataflowErrorsTest extends InterpreterTest {
           |""".stripMargin
       eval(code)
       consumeOut shouldEqual List(
-        "(Error: (Syntax_Error_Data 'Unrecognized token.'))",
-        "(Syntax_Error_Data 'Unrecognized token.')"
+        "(Error: (Syntax_Error.Error 'Unrecognized token.'))",
+        "(Syntax_Error.Error 'Unrecognized token.')"
       )
     }
   }