+++ std-bits/table/src/main/java/org/enso/table/data/table/join/Between.java
@@ -1,0 +1,5 @@
+package org.enso.table.data.table.join;
+
+import org.enso.table.data.table.Column;
+
+public record Between(Column left, Column rightLower, Column rightUpper) implements JoinCondition {}